Add files via upload
A sleek and responsive stopwatch built using HTML, CSS, and JavaScript. This project features intuitive controls for start, stop, and reset functionalities, along with a visually appealing design. It’s perfect for tracking time accurately and efficiently with a user-friendly interface. Explore this project to see the magic of simple web technologies combined to create a fully functional stopwatch. Ideal for those looking to learn or demonstrate basic web development skills.